Module Glue.MinimalSemiLattice

Parameters

module P : sig ... end

Signature

type property = P.property

val leq_join : property -> property -> property

leq_join p q must compute the join of p and q. If the result is logically equal to q, then q itself must be returned. Thus, we have leq_join p q == q if and only if leq p q holds.