Module Fix.Prop

Prop offers a number of ready-made implementations of the signature PROPERTY.

module Boolean : sig ... end

The lattice of Booleans.

The lattice of options.

This ordering is not a lattice.

Although the code is polymorphic in the type of elements, it must still be packaged as a functor, because property cannot be a parameterized type.

module Option (X : sig ... end) : sig ... end

The lattice of sets.

module Set (X : sig ... end) : sig ... end