Module Prop.Boolean

The lattice of Booleans.

The Boolean lattice. The ordering is false <= true.

type property = bool

val bottom : property

val equal : property -> property -> bool

val is_maximal : property -> bool

val leq : bool -> bool -> bool

val join : bool -> bool -> bool

val leq_join : bool -> bool -> bool

leq_join p q must compute the join of p and q. If the result is logically equal to q, then q itself must be returned. Thus, we have leq_join p q == q if and only if leq p q holds.